Made it a bit more robust and added no-clobber option.
git-svn-id: https://svn.code.sf.net/p/ngsgenomelibs/code/trunk@233 2b8d95ea-eb58-0410-8a55-29fbd2c437c3

+++ perl_utils_usc/buildMySqlTablesByChrom.pl
@@ -5,48 +5,53 @@ use File::Basename;
 use Getopt::Long;
 use File::Temp qw/ tempfile /;
 
-my $USAGE = "buildMySqlTablesByChrom.pl --templateFile schema.sql [--pipeConverterScript pipe.pl] [--db cr] [--replaceStr TEMPLATE_CHR] table1_chr1.txt table1_chr1.txt ...";
+my $USAGE = "buildMySqlTablesByChrom.pl [--noclobberTable|--templateFile schema.sql] [--pipeConverterScript pipe.pl] [--db cr] [--replaceStr TEMPLATE_CHR] table1_chr1.txt table1_chr1.txt ...";
 
 my $delim = "\t";
 my $templateFn = 0;
 my $replaceStr = "TEMPLATE_CHR";
 my $pipe = 0;
+my $clobberTable = 1;
 my $db = "cr";
-GetOptions ('pipeConverterScript=s'=>\$pipe, 'templateFile=s' => \$templateFn, 'replaceStr=s' => \$replaceStr) || die "$USAGE\n";
-die "$USAGE\n" unless ($templateFn && (@ARGV>0));
+GetOptions ('clobberTable!'=>\$clobberTable, 'pipeConverterScript=s'=>\$pipe, 'templateFile=s' => \$templateFn, 'replaceStr=s' => \$replaceStr) || die "$USAGE\n";
+
+print STDERR "clobber=$clobberTable\targv=" . scalar(@ARGV) . "\n";
+die "$USAGE\n" unless ((!$clobberTable || $templateFn) && (@ARGV>0));
 
 foreach my $tabfn (@ARGV)
 {
     my ($name, $path, $suf) = fileparse($tabfn, qr/\.[^.]*/);
 
-	die "Can't read $templateFn\n" unless (open(TF,$templateFn));
-	my @templateLines = ();
-	while (my $line = <TF>)
-	{
-		push(@templateLines,$line);
-	}
-	close(TF);
-
-	my $tmpfn = makeTableFile(\@templateLines, $replaceStr, $name);
+ 	if ($clobberTable)
+ 	{
+		die "Can't read $templateFn\n" unless (open(TF,$templateFn));
+		my @templateLines = ();
+		while (my $line = <TF>)
+		{
+			push(@templateLines,$line);
+		}
+		close(TF);
 	
-	runCmd("cat ${tmpfn} | mysql ${db}");
-	unlink($tmpfn);
-	#runCmd("echo \"drop table ${name};\" | mysql ${db}"); #----REMOVE
-
+		my $tmpfn = makeTableFile(\@templateLines, $replaceStr, $name);
+		
+		runCmd("cat ${tmpfn} | mysql ${db}");
+		unlink($tmpfn);
+		#runCmd("echo \"drop table ${name};\" | mysql ${db}"); #----REMOVE
+	}
 
 	# Do we need to pipe the file through a converter?  Since the file has to have the same name as the 
 	# table, we have to move it.
 	if ($pipe)
 	{
-		my $origTabfn = "${tabfn}.orig";
-		runCmd("mv $tabfn $origTabfn");
-		
-		my $cmd = "${pipe} < $origTabfn > $tabfn";
+		my $tabfnNew = "${name}.converted.txt";
+		my $cmd = "${pipe} < $tabfn > $tabfnNew";
 		runCmd($cmd);
+		$tabfn = $tabfnNew;
 	}
 
-	
-	my $cmd = "mysqlimport --use-threads=2 --lock-tables --replace --delete --force --local --verbose ${db} ${tabfn}";
+
+	my $deleteSec = ($clobberTable) ? " --delete " : "";	
+	my $cmd = "mysqlimport ${deleteSec} --use-threads=2 --lock-tables --replace --force --local --verbose ${db} ${tabfn}";
 	runCmd($cmd);
 }
